5 February, 2001 - Jay Larson <larson@mcs.anl.gov>
Added numerous query and local-to-global process mapping services to
the MCTWorld data type.  These are:

NumComponents - The number of components in MCTWorld
ComponentNumProcs - The number of processes a component covers
ComponentToWorldRank - Local to global process mapping
ComponentRootRank - Returns global address of a component's root.

These have been compiled/built into a library and introduced no
complications to building ut-mct on the IRIX64 platform.  That
said, the functions remain to be application-tested.