6 February 2001 - Jay Larson <larson@mcs.anl.gov>
Modified the GlobalSegMap to remove the hard-wired %lsize component.
Now, GlobalSegMap_lsize() is a computed function.  Also, added a new
query function called GlobalStorage(), which returns the total length
of a vector needed to store the distributed data on 1 process,
including halo points (if present).

The change to GlobalSegMap_lsize() included a change in its interface,
and this change had to be included in calls to this function from the
Router module.