13 June 2001 - J. Larson <larson@mcs.anl.gov>
Changes to the GeneralGrid init() procedures.  In the past, init()
has initialized List (see m_List in mpeu for details) components of
the GeneralGrid as blank if no initialization data is provided.  This
has caused problems on the AIX platform in the blocking GeneralGrid
send() and recv() operations.

This check-in is designed to eliminate this problem.