14 June 2001 - Jay Larson
Checked in patch to a bug in the routine nlseg_() of m_GlobalSegMap.
This bug is explained in the "REVISION HISTORY" of this routine.  The
bug itself was discovered by Everest Ong.

The other change is a slight modification of the abort condition at
the top of the routine GM_scatter_() of m_AttrVectComms.  Now, the
process ID on which the error is detected is included in the message
written to the device stderr.