18 June 2001 - Jay Larson <larson@mcs.anl.gov>
Added a nullify_() method to the List class.  This was added in
response to problems under AIX regarding uninitialized pointers.

Now, if a user is creating an object with List components that
are not going to be initialized, List_nullify() should be invoked.

Invoking List_nullify guarantees unambiguous/incorrect results from
the f90 intrinsic associated().