February 14, 2002 14:20 - Jay Larson <larson@mcs.anl.gov>
Two modifications to the AttrVect:

1) Bug fix to the routines exportRList() and exportIList().
The routine List_copy was not referenced by a use statement,
which created problems at link time for the unit tester.

2) Addition of the CHARACTER functions exportRListToChar()
and exportIListToChar(), which allows the mct user to write
out AttrVect rList and iList components without explicit
reference to the list character buffer.  This has been tested
with success in ut-mct.