6 June 2002 J. Larson
Modified routine LocalReduceReals_() in m_AttrVect to work around
compiler bug in the SGI v7.3 f90 compiler.

In the previous version of the routine LocalReduceReals_() in m_AttrVect, the output
AttrVect was initialized using a call to init_(), where the real attribute list argument
was a string from exportRListToChar().  A bit too OO for the compiler.

The new version sports a nullified integer Attribute List, and is used it along with
inAv%rList to invoke initl_().  This works on the SGI.