9 July 2002 - J. Larson <larson@mcs.anl.gov>
Added the initial implementation of the Merge facility
(m_Merge.F90 is a new file), and made changes to SRCS_F90.mk
and Makefile.dependency to reflect this addition.

Currently, the merge works on a fairly narrow situation where
the output AttrVect shares the same attributes with all of the
input AttrVect arguments.  This assumption is fairly safe for the
climate applications we currently support.  Future versions of the
merge will be more sophisticated (don't hold your breath...I'm talking
version 2 here).