29 July 2002 - J. Larson
This commit reflects documentation added to the String and List
modules.  The bulk of the changes are the insertion of complete
protex-compliant (tested) prologues to these modules for future
incorporation as an appendix to the MCT API document.  There were
also some argument checks added to the getrange() routine in the
List module.