20 August 2002 - J. Larson
Port to the Fujitsu-VPP.

* Added Make configurations (Makefile.conf.Fujitsu-VPP) in both
the mpeu and mpeu/mpi0 directories.

Changed the configuration scripts configure and mpi0/configure
to reflect this new platform.

m_FileResolv was changed to correct a bug caught by the Fujitsu
compiler.  The compiler caught this in its semi-strict mode (for
those who don't know, the Fujitsu compiler's strictest syntax
parsing mode is about as rigorous as flint).

m_inpak90 was modified to lower the character buffer size used to
store an input resource file down to the maximum allowed on the
Fujitsu platform.

m_ioutil was modified to make explicit the connection between our
version of flush() and what is available on the VPP.