20 August 2002 - J. Larson
Port to the Fujitsu-VPP platform.

The configure script was modified to recognize this new platform.

A set of make configurations (in the file Makefile.conf.Fujitsu-VPP)
was added.

The module m_AttrVectReduce was modified to define explicitly in the
module declaration section the values of the public data members
AttrVect_SUM, AttrVect_MIN, and AttrVect_MAX.  Previously, the mpeu
module m_mpif90 was used in the declaration section, and the
aforementioned public data members were set from their MPI operation
counterparts (i.e., MPI_SUM, MPI_MIN, and MPI_MAX).  This simply would
not work on the Fujitsu--hence the now explicit values.  This change
also affected the GlobalReduce() routine as the AttrVect reduction
operation values were no longer set directly from their MPI
counterparts.  Bottom line:  a 'select case' block was added to handle
the various reduction operations and relate them to their MPI
counterparts for the call to MPI_ALLREDUCE().

The module m_SparseMatrixComms had a bug in its broadcast routine.
The local process ID had not been set (which was detected by the
Fujitsu compiler).  This bug was previously undetected because the
routine has (to date) not been tested.