26 September, 2002 - J. Larson
This commit reflects the following changes:

To the Module m_ConvertMaps:

Completion of the GlobalSegMap->GlobalMap conversion routine,
and generally cleaned up documentation.

To the Module m_ExchangeMaps:

Removal of two empty APIs that output GlobalMap datatypes.  This
is in keeping with MCT's policy of supplying only GlobalSegMap-
based APIs for most of its high-level services.  The API that accepts
a GlobalMap for input was finally implemented.  This implementation
creates a new dependency on the module m_ConvertMaps.

To the file Makefile.dependency:

Added the new dependency of m_ExchangeMaps on m_ConvertMaps mentioned
in the log entry for m_ExchangeMaps.